feat: homepage MVP (#190)
I applied all the older styles from our svelte concept to the new homepage.

The footer component, as well as the dark mode color structure, should be implemented separately (using a new branch).

Feel free to suggest visual changes, as I can currently only test the website on a 13-inch screen. In addition, the content of the navbar is still missing, which can lead to space bugs.

Co-authored-by: Victor Homic <87198856+3x071c@users.noreply.github.com>